requirements.txt
fastapi
Pillow
python-multipart
torch
torchvision
uvicorn